# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.4.0",
        "title": "Aimedis API",
        "description": "Manage the Api",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "/api/v4",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/step/GetStep/{house_id}": {
            "parameters": [
                {
                    "name": "token",
                    "in": "header",
                    "description": "JWT token User"
                },
                {
                    "name": "house_id",
                    "in": "path",
                    "description": "Id of house"
                }
            ],
            "get": {
                "tags": [
                    "Patientflow"
                ],
                "summary": "Get information from house_id ",
                  "responses": {
                    "200": {
                        "description": "information is found"
                    }
                }
            }
        },
...
        "/step/AddStep": {
            "parameters": [
                {
                    "name": "token",
                    "in": "header",
                    "description": "JWT token User"
                }
            ],
            "post": {
                "tags": [
                    "Patientflow"
                ],
                "parameters": [
                    {
                        "name": "AddStep",
                        "in": "body",
                        "description": "patientflow",
                        "schema": {
                            "$ref": "#/definitions/Addstep"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "step added succefully",
                        "schema": {
                            "$ref": "#/definitions/Addstep"
                        }
                    },
                    "500": {
                        "description": "Something happened wrong"
                    }
                }
            }
        },
        "/step/Case_numbers/{step_id}": {
            "parameters": [
                {
                    "name": "token",
                    "in": "header",
                    "description": "JWT token User"
                },
                {
                    "name": "step_id",
                    "in": "path",
                    "description": "Step id"
                }
            ],
            "post": {
                "tags": [
                    "Patientflow"
                ],
                "parameters": [
                    {
                        "name": "Case_numbers",
                        "in": "body",
                        "description": "patientflow",
                        "schema": {
                            "$ref": "#/definitions/StepAddCase"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "case added succefully",
                        "schema": {
                            "$ref": "#/definitions/Case_numbers"
                        }
                    },
                    "500": {
                        "description": "Something happened wrong"
                    }
                }
            }
        },
...
        "/step/AddStep/{step_id}":{
            "parameters": [
                {
                    "name": "token",
                    "in": "header",
                    "description": "JWT token User"
                },
                {
                    "name": "step_id",
                    "in": "path",
                    "required": true,
                    "description": "ID",
                    "type": "string"
                }
            ],
            "delete": {
                "summary": "Delete with given ID",
                "tags": [
                    "Patientflow"
                ],
                "responses": {
                    "200": {
                        "description": "step deleted",
                        "schema": {
                            "$ref": "#/definitions/AddStep"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Patientflow"
                ],
                "parameters": [
                    {
                        "name": "Patientflow",
                        "in": "body",
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Addstep"
                        }
                    }
                ],
                "summary": "Update with given ID",
                "responses": {
                    "200": {
                        "description": "step updated"
                    }
                }
            }
        },
        "/step/Case_numbers/{step_id}/{case_id}":{
            "parameters": [
                {
                    "name": "token",
                    "in": "header",
                    "description": "JWT token User"
                },
                {
                    "name": "step_id",
                    "in": "path",
                    "required": true,
                    "description": "ID",
                    "type": "string"
                },
                {
                    "name": "case_id",
                    "in": "path",
                    "required": true,
                    "description": "ID",
                    "type": "string"
                }
                
            ],
            "delete": {
                "summary": "Delete with given ID",
                "tags": [
                    "Patientflow"
                ],
                "responses": {
                    "200": {
                        "description": "case is deleted"
                    }
                }
            },
            "put": {
                "tags": [
                    "Patientflow"
                ],
                "parameters": [
                    {
                        "name": "Patientflow",
                        "in": "body",
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/StepAddCase"
                        }
                    }
                ],
                "summary": "Update with given ID",
                "responses": {
                    "200": {
                        "description": "case is updated"
                    }
                }
            }
        },
...
        "/step/id_numbers/cases/{fromstep_id}/{tostep_id}":{
            "parameters": [
                {
                    "name": "token",
                    "in": "header",
                    "description": "JWT token User"
                },
                {
                    "name": "fromstep_id",
                    "in": "path",
                    "required": true,
                    "description": "ID",
                    "type": "string"
                },
                {
                    "name": "tostep_id",
                    "in": "path",
                    "required": true,
                    "description": "ID",
                    "type": "string"
                }
                
            ],
            
            "put": {
                "tags": [
                    "Patientflow"
                ],
                "parameters": [
                    {
                        "name": "Patientflow",
                        "in": "body",
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Addstep"
                        }
                    }
                ],
                "summary": "update with given iD",
                "responses": {
                    "200": {
                        "description":"updated successfully"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Addstep": {
            "required": ["step_name","house_id" ],
            "properties": {
                "step_name": {
                    "type": "string",
                    "uniqueItems": false
                },
                "step_order": {
                    "type": "string",
                    "uniqueItems": false
                },
                "house_id": {
                    "type": "string",
                    "uniqueItems": false
                }
            }
        },
        "Case_numbers": {
            "required": ["step_name","house_id" ],
            "properties": {
                "step_name": {
                    "type": "string",
                    "uniqueItems": false
                },
                "step_order": {
                    "type": "string",
                    "uniqueItems": false
                },
                "house_id": {
                    "type": "string",
                    "uniqueItems": false
                }
            }
        },
        "StepAddCase":{
            "required": ["case_id"],
            "properties": {
                "case_id": {
                    "type": "string",
                    "uniqueItems": false
                }
            }
        },
...
    }
    }